move kube api utils to common
