+++ tests/tests_deployment/test_loki_deployment.py
@@ -1,47 +1,14 @@
 import pytest
 
-import socket
-
 import six.moves.urllib.request as urllib_request
 
-from kubernetes import config
-from kubernetes.client import Configuration
-from kubernetes.client.api import core_v1_api
-from kubernetes.stream import portforward
-
+from tests.common.kube_api import kubernetes_port_forward
 
 LOKI_BACKEND_PORT = 3100
 LOKI_BACKEND_POD_LABELS = {
     "app.kubernetes.io/instance": "nebari-loki",
     "app.kubernetes.io/component": "backend"
 }
-
-
-def kubernetes_port_forward(pod_labels, port, namespace="dev"):
-    """Given pod labels and port, finds the pod name and port forwards to
-    the given port.
-    """
-    config.load_kube_config()
-    Configuration.set_default(Configuration.get_default_copy())
-    core_v1 = core_v1_api.CoreV1Api()
-
-    pods = core_v1.list_namespaced_pod(
-        namespace=namespace,
-        label_selector=pod_labels
-    )
-    assert pods.items
-    pod = pods.items[0]
-    pod_name = pod.metadata.name
-
-    def kubernetes_create_connection(address, *args, **kwargs):
-        pf = portforward(
-            core_v1.connect_get_namespaced_pod_portforward,
-            pod_name, namespace, ports=str(port)
-        )
-        return pf.socket(port)
-
-    socket.create_connection = kubernetes_create_connection
-    return pod
 
 
 @pytest.fixture(scope="module")
@@ -64,6 +31,9 @@ def port_forward():
     ),
 )
 def test_loki_endpoint(endpoint_path, port_forward):
+    """This will hit some endpoints in the loki API and verify that we
+    get a 200 status code.
+    """
     pod_name = port_forward.metadata.name
     url = f'http://{pod_name}.pod.dev.kubernetes:{LOKI_BACKEND_PORT}/{endpoint_path}'
     response = urllib_request.urlopen(url)
